chore: Moved and renamed some files
