+++ taletime/lib/player/widgets/story_image.dart
